@@ -1,11 +1,19 @@
 require 'spec_helper'
 
 describe Codeqa::RunnerDecorator do
-  it "should run the matching_checkers" do
-    source = source_with("def syntax_error", "ruby.rb")
+  it 'should run provide the errors if the checker failed' do
+    source = source_with('def syntax_error', 'ruby.rb')
     runner = Codeqa::Runner.run(source)
-    runner.should_not be_success
+    expect(runner.success?).to be false
     decorator = Codeqa::RunnerDecorator.new(runner)
-    decorator.to_s.should match(/ruby syntax/)
+    expect(decorator.to_s).to match(/syntax/)
   end
+  it 'should run list the ran checkers' do
+    source = source_with('def foo; end', 'ruby.rb')
+    runner = Codeqa::Runner.run(source)
+    expect(runner.success?).to be true
+    decorator = Codeqa::RunnerDecorator.new(runner)
+    expect(decorator.to_s).to match(/Passed tests.+strange chars/)
+  end
+
 end
