@@ -8,11 +8,11 @@ module Codeqa
       end
 
       def name
-        "erb html"
+        'erb html'
       end
 
       def hint
-        "The html I see after removing the erb stuff is not valid (find the unclosed tags and attributes)."
+        'The html I see after removing the erb stuff is not valid (find the unclosed tags and attributes).'
       end
 
       def check
@@ -24,18 +24,17 @@ module Codeqa
           end # IO.popen
         end # Tempfile
 
-        if result
-          errors.add(nil, html)
-          errors.add(nil, result)
-        end
+        return unless result
+        errors.add(nil, html)
+        errors.add(nil, result)
       end
 
       def html
         @html ||= begin
                     html = FakeERB.new(sourcefile.content.gsub('<%=', '<%')).result
-                    html = html.force_encoding("UTF-8") if html.respond_to?(:force_encoding)
+                    html = html.force_encoding('UTF-8') if html.respond_to?(:force_encoding)
                     html.gsub(%r{<script[ >].*?</script>|<style[ >].*?</style>}m,
-                              "<!--removed script/style tag-->")
+                              '<!--removed script/style tag-->')
                   end
       end
     end
