@@ -7,17 +7,17 @@ module Codeqa
       end
 
       def name
-        "ruby syntax"
+        'ruby syntax'
       end
 
       def hint
-        "Ruby can not parse the file, please check it for syntax errors."
+        'Ruby can not parse the file, please check it for syntax errors.'
       end
 
       def check
         with_existing_file do |filename|
           command = "/usr/bin/env ruby -c '#{filename}' 1>/dev/null 2>/dev/null"
-          errors.add(nil, "Ruby syntax error") unless system(command)
+          errors.add(nil, 'Ruby syntax error') unless system(command)
         end
       end
     end
