@@ -1,3 +1,3 @@
 module Codeqa
-  VERSION = "0.2.6"
+  VERSION = '0.3.0'
 end
