@@ -1,13 +1,8 @@
-# This file was generated by the `rspec --init` command. Conventionally, all
-# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
-# Require this file using `require "spec_helper"` to ensure that it is only
-# loaded once.
-#
-# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+require 'simplecov'
+SimpleCov.start
 
 require 'codeqa'
 RSpec.configure do |config|
-  config.treat_symbols_as_metadata_keys_with_true_values = true
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
 
@@ -16,6 +11,28 @@ RSpec.configure do |config|
   # the seed, which is printed after each run.
   #     --seed 1234
   config.order = 'random'
+
+  # override some config settings
+  config.before(:all) do
+    load_test_config
+  end
+end
+
+Dir['./spec/support/**/*.rb'].sort.each{ |f| require f }
+
+def load_test_config
+  Codeqa.configure do |c|
+    c.excludes = ['vendor/**/*']
+    c.enabled_checker = %w(CheckStrangeChars
+                           CheckUtf8Encoding
+                           CheckConflict
+                           CheckPry
+                           RubocopLint)
+    c.rubocop_formatter_cops = %w(EmptyLinesAroundBody
+                                  EmptyLines
+                                  TrailingWhitespace)
+    c.erb_engine = 'erb'
+  end
 end
 
 def check_with(klass, source)
